Add voice obfuscation voice to non-humans, add examine text for obfuscation
