+++ Content.Client/_ES/Chat/Obfuscation/ESVoiceObfuscatorSystem.cs
@@ -1,0 +1,6 @@
+using Content.Shared._ES.Chat.Obfuscation;
+
+namespace Content.Client._ES.Chat.Obfuscation;
+
+/// <inheritdoc/>
+public sealed class ESVoiceObfuscatorSystem : ESSharedVoiceObfuscatorSystem;
