+++ Content.Server/_ES/Chat/Obfuscation/ESVoiceObfuscatorSystem.cs
@@ -1,24 +1,21 @@
-using Content.Server._ES.Chat.Obfuscation.Components;
-using Content.Server.Humanoid;
+using Content.Shared._ES.Chat.Obfuscation;
+using Content.Shared._ES.Chat.Obfuscation.Components;
 using Content.Shared.Chat;
 using Content.Shared.Clothing.EntitySystems;
-using Content.Shared.Humanoid;
-using Content.Shared.IdentityManagement.Components;
 using Content.Shared.Inventory;
 
 namespace Content.Server._ES.Chat.Obfuscation;
 
-/// <summary>
-/// This handles <see cref="ESVoiceObfuscatorComponent"/>
-/// </summary>
-public sealed class ESVoiceObfuscatorSystem : EntitySystem
+/// <inheritdoc/>
+public sealed class ESVoiceObfuscatorSystem : ESSharedVoiceObfuscatorSystem
 {
-    [Dependency] private readonly HumanoidAppearanceSystem _humanoidAppearance = default!;
     [Dependency] private readonly MaskSystem _mask = default!;
 
     /// <inheritdoc/>
     public override void Initialize()
     {
+        base.Initialize();
+
         SubscribeLocalEvent<ESVoiceObfuscatorComponent, InventoryRelayedEvent<TransformSpeakerNameEvent>>(OnTransformSpeakerName);
     }
 
@@ -28,22 +25,5 @@ public sealed class ESVoiceObfuscatorSystem : EntitySystem
             return;
 
         args.Args.VoiceName = GetObfuscatedVoice(args.Owner);
-    }
-
-    private string GetObfuscatedVoice(Entity<HumanoidAppearanceComponent?> ent)
-    {
-        // They need to have this component.
-        if (!Resolve(ent, ref ent.Comp))
-            return string.Empty;
-
-        var species = ent.Comp.Species;
-        var age = ent.Comp.Age;
-
-        var name = Name(ent);
-        var gender = ent.Comp.Gender;
-        var ageRepresentation = _humanoidAppearance.GetAgeRepresentation(species, age);
-        var identityRepresentation = new IdentityRepresentation(name, gender, ageRepresentation);
-
-        return identityRepresentation.ToStringUnknown();
     }
 }
