+++ Content.Shared/_ES/Chat/Obfuscation/Components/ESGenericVoiceComponent.cs
@@ -1,0 +1,17 @@
+using Robust.Shared.GameStates;
+
+namespace Content.Shared._ES.Chat.Obfuscation.Components;
+
+/// <summary>
+/// Component that provides a "general" voice to use when the entity's voice is obfuscated via <see cref="ESVoiceObfuscatorComponent"/>.
+/// </summary>
+[RegisterComponent, NetworkedComponent]
+[Access(typeof(ESSharedVoiceObfuscatorSystem))]
+public sealed partial class ESGenericVoiceComponent : Component
+{
+    /// <summary>
+    /// String that will be used as the voice name
+    /// </summary>
+    [DataField(required: true)]
+    public LocId Voice;
+}
