+++ Content.Shared/_ES/Chat/Obfuscation/Components/ESVoiceObfuscatorComponent.cs
@@ -1,9 +1,11 @@
-namespace Content.Server._ES.Chat.Obfuscation.Components;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared._ES.Chat.Obfuscation.Components;
 
 /// <summary>
 /// This is used for clothing items which hide the name of someone speaking.
 /// This version replaces it with a generic descriptor of their appearance.
 /// </summary>
-[RegisterComponent]
-[Access(typeof(ESVoiceObfuscatorSystem))]
+[RegisterComponent, NetworkedComponent]
+[Access(typeof(ESSharedVoiceObfuscatorSystem))]
 public sealed partial class ESVoiceObfuscatorComponent : Component;
