+++ Resources/Locale/en-US/_ES/chat/voice-obfuscator.ftl
@@ -1,0 +1,3 @@
+es-voice-obfuscator-examine = Wearing this will [color=crimson]hide your identity[/color] when speaking.
+
+es-voice-obfuscator-name-replace-monkey = monkey
