+++ Resources/Prototypes/Entities/Mobs/NPCs/animals.yml
@@ -1478,6 +1478,13 @@
         Blunt: 1
     clumsySound:
       path: /Audio/Animals/monkey_scream.ogg
+# ES START
+  - type: Grammar # adding this in so that a named monkey like Pun-Pun always displays the same as a normal monkey
+    attributes:
+      proper: true
+  - type: ESGenericVoice
+    voice: es-voice-obfuscator-name-replace-monkey
+# ES END
 
 
 - type: entity
